+++ infra/executor/src/test/java/org/apache/shardingsphere/infra/executor/sql/execute/result/query/impl/driver/jdbc/type/stream/JDBCStreamQueryResultTest.java
@@ -115,7 +115,7 @@ class JDBCStreamQueryResultTest {
     @Test
     void assertGetValueByBytes() throws SQLException {
         ResultSet resultSet = mock(ResultSet.class);
-        byte[] value = {1};
+        byte[] value = {(byte) 1};
         when(resultSet.getBytes(1)).thenReturn(value);
         assertThat(new JDBCStreamQueryResult(resultSet).getValue(1, byte[].class), is(value));
     }
